Move legacy-only constants out of src/constants
Keep src/utils/constants.py focused on project-level constants used by the current code path, move legacy constants to legacy/utils/legacy_constants.py, and rewire legacy/example/wikipedia imports accordingly.
